fix invalid type casting of size_t
This patch fixes invalid type casting of size_t. There was no problem
of 32bit system but it can be faulty from 64bit compiler.

+++ dp_buffer.c
@@ -12,7 +12,7 @@
 #include "dp.h"
 
 unsigned int dp_buffer_alloc(int fd, int width, int height, int bpp,
-				unsigned int *handle, int *pitch, int *size)
+				unsigned int *handle, int *pitch, size_t *size)
 {
 
 	struct drm_mode_create_dumb args;
